xilfpga:src:MISRA-C:Use of mixed mode arithmetic
This patch fixes below MISRA-C violation in Xilfpga
Use of mixed mode arithmetic
MISRA-C:2012/AMD1 R.10.1,R.10.2,R.10.3,R.10.4


Acked-for-series: Nava kishore Manne <nava.manne@xilinx.com>

+++ lib/sw_services/xilfpga/src/interface/zynqmp/xilfpga_pcap.c
@@ -947,7 +947,7 @@ static u32 XFpga_SecureBitstreamDdrLoad(UINTPTR BitstreamAddr, UINTPTR KeyAddr,
 				XSECURE_AUTH_CERT_MIN_SIZE/(u32)XSECURE_WORD_LEN);
 		/*Verify Spk */
 		Status = XSecure_VerifySpk(AcBuf, ImageInfo->EfuseRsaenable);
-		if (Status != XST_SUCCESS) {
+		if (Status != (u32)XST_SUCCESS) {
 			Status = XFPGA_PCAP_UPDATE_ERR(
 					XFPGA_ERROR_DDR_AUTH_VERIFY_SPK,
 					Status);
@@ -958,7 +958,7 @@ static u32 XFpga_SecureBitstreamDdrLoad(UINTPTR BitstreamAddr, UINTPTR KeyAddr,
 		Status = XSecure_PartitionAuthentication(&CsuDma, (u8 *)BitAddr,
 							 PL_PARTATION_SIZE,
 							(u8 *)(UINTPTR)AcBuf);
-		if (Status != XST_SUCCESS) {
+		if (Status != (u32)XST_SUCCESS) {
 			Status = XFPGA_PCAP_UPDATE_ERR(
 					XFPGA_ERROR_DDR_AUTH_PARTITION,
 					Status);
@@ -990,7 +990,7 @@ static u32 XFpga_SecureBitstreamDdrLoad(UINTPTR BitstreamAddr, UINTPTR KeyAddr,
 				XSECURE_AUTH_CERT_MIN_SIZE/(u32)XSECURE_WORD_LEN);
 		/*Verify Spk */
 		Status = XSecure_VerifySpk(AcBuf, ImageInfo->EfuseRsaenable);
-		if (Status != XST_SUCCESS) {
+		if (Status != (u32)XST_SUCCESS) {
 			Status = XFPGA_PCAP_UPDATE_ERR(
 					XFPGA_ERROR_DDR_AUTH_VERIFY_SPK,
 					Status);
@@ -1001,7 +1001,7 @@ static u32 XFpga_SecureBitstreamDdrLoad(UINTPTR BitstreamAddr, UINTPTR KeyAddr,
 		Status = XSecure_PartitionAuthentication(&CsuDma, (u8 *)BitAddr,
 							 RemaningBytes,
 							(u8 *)(UINTPTR)AcBuf);
-		if (Status != XST_SUCCESS) {
+		if (Status != (u32)XST_SUCCESS) {
 			Status = XFPGA_PCAP_UPDATE_ERR(
 					XFPGA_ERROR_DDR_AUTH_PARTITION,
 					Status);
@@ -1162,7 +1162,7 @@ static u32 XFpga_AuthPlChunksOcm(XFpga *InstancePtr, u32 Size)
 		XSECURE_AUTH_CERT_MIN_SIZE/(u32)XSECURE_WORD_LEN);
 	/*Verify Spk */
 	Status = XSecure_VerifySpk(AcBuf, ImageInfo->EfuseRsaenable);
-	if (Status != XST_SUCCESS) {
+	if (Status != (u32)XST_SUCCESS) {
 		Status = XFPGA_PCAP_UPDATE_ERR(
 				XFPGA_ERROR_OCM_AUTH_VERIFY_SPK,
 				Status);
